requirements.txt
# Flask
click==8.1.7
Flask==3.1.1
Werkzeug==3.1.3
gunicorn==23.0.0
six==1.16.0
python-dotenv==1.0.1
python-dateutil==2.9.0.post0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
Flask-Migrate==4.0.7
sqlalchemy-json==0.7.0
cloud-sql-python-connector==1.14.0
pg8000==1.31.2

# API and Models
Flask-RESTful==0.3.10
marshmallow==3.23.1
marshmallow-sqlalchemy==1.1.0
flask-marshmallow==1.2.1
flask-apispec==0.11.4

# Security
Flask-Cors==6.0.0
flask-talisman==1.1.0
PyJWT==2.10.1
flask-oidc==2.2.2

# Requests
requests==2.32.4

# Error Reporting
sentry-sdk[flask]==2.20.0

# Okta
okta==2.9.8

# Plugins
pluggy==1.5.0

# FastAPI (for migration to FastAPI)
fastapi[standard]==0.116.1
pydantic==2.11.7
python-multipart==0.0.20

# Test - TODO: Move test packages to separate file
tox==4.23.2
# Lint
ruff==0.8.0
# Typing
mypy==1.13.0
types-google-cloud-ndb==2.3.0.20241103
types-Flask-Cors==5.0.0.20240902
types-Flask-Migrate==4.0.0.20240311
types-python-dateutil==2.9.0.20241003
types-requests==2.32.0.20241016
